azure_documenter/requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-web
azure-mgmt-keyvault
azure-mgmt-authorization
azure-mgmt-policyinsights
azure-mgmt-security
azure-mgmt-advisor
azure-mgmt-consumption
jinja2
markdown
graphviz
openai
python-dotenv
pandas 